requirements.txt
flask==3.1.3
playwright==1.58.0
pyngrok==7.5.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
